requirements.txt
python-dotenv
requests
redis

fastapi==0.110.0
uvicorn[standard]==0.29.

kafka-python